:sparkles: Add python script
These scripts generates a window with TKinter in order to select a font amongst the locally installed one, and let you preview the rendered character pixelized onto a given size. Then, you choose a file full of symbols, and every character in it is converted using chosen font and size.